+++ examples/audit-test/src/integration-test/groovy/test/AuditUpdateSpec.groovy
@@ -49,12 +49,6 @@ class AuditUpdateSpec extends Specification {
                 heliport.save(flush: true, failOnError: true)
             }
         }
-        EntityInSecondDatastore.withNewTransaction {
-            AuditLogContext.withoutAuditLog {
-                EntityInSecondDatastore.where {}.deleteAll()
-                new EntityInSecondDatastore(name: "name", someIntegerProperty: 1).save(flush: true, failOnError: true)
-            }
-        }
         AuditTrail.withNewTransaction {
             AuditTrail.where {}.deleteAll()
             assert AuditTrail.count() == 0
@@ -255,125 +249,5 @@ class AuditUpdateSpec extends Specification {
         ['name', 'famous'].each { name ->
             assert events.find { it.propertyName == name }, "${name} was not logged"
         }
-    }
-
-    void "Test rollback behaviour"() {
-        when:
-        Author.withNewTransaction { TransactionStatus transactionStatus ->
-            def author = Author.findByName("Aaron")
-            author.age = 1
-            Author.withSession { Session session ->
-                session.flush()
-            }
-            transactionStatus.setRollbackOnly()
-        }
-
-        then:
-        Author.withNewTransaction { TransactionStatus transactionStatus ->
-            Author.findByName("Aaron").age == 37
-        }
-        AuditTrail.withNewTransaction {
-            AuditTrail.list() == []
-        }
-
-        when:
-        Author.withNewTransaction {
-            def author = Author.findByName("Aaron")
-            author.age = 3
-        }
-
-        then:
-        Author.withNewTransaction {
-            Author.findByName("Aaron").age == 3
-        }
-        AuditTrail.withNewTransaction {
-            AuditTrail.count
-        } == 1
-    }
-
-    void "Test nested transactions"() {
-        when:
-        Author.withNewTransaction { TransactionStatus transactionStatus ->
-            Author.findByName("Aaron").age = 1
-            Author.withSession { Session session ->
-                session.flush()
-            }
-            Book.withNewTransaction { TransactionStatus transactionStatus2 ->
-                Book.findByTitle("Hunger Games").pages = 401
-            }
-            transactionStatus.setRollbackOnly()
-        }
-
-        then:
-        Author.withNewTransaction {
-            Author.findByName("Aaron")
-        }.age == 37 // rolled back change -> original value
-        Book.withNewTransaction {
-            Book.findByTitle("Hunger Games")
-        }.pages == 401 // committed change -> new value
-        AuditTrail.withNewTransaction {
-            AuditTrail.count
-        } == 1 // AuditTrail only for committed change
-        AuditTrail.withNewTransaction {
-            AuditTrail.list()[0]
-        }.newValue == "401"
-    }
-
-    void "test nested transactions different datastores: rollback outer"() {
-        expect:
-        AuditTrail.withNewTransaction {
-            AuditTrail.list().collect { it.className }.unique()
-        } == []
-
-        when:
-        Author.withNewTransaction { TransactionStatus transactionStatus ->
-            EntityInSecondDatastore.withNewTransaction {
-                new Author(name: "name2", age: 12, famous: true).save(flush: true, failOnError: true)
-                new EntityInSecondDatastore(name: "name2", someIntegerProperty: 1).save(flush: true, failOnError: true)
-                // Commit new EntityInSecondDatastore
-            }
-            // Rollback new Author
-            transactionStatus.setRollbackOnly()
-        }
-
-        then:
-        Author.withNewTransaction {
-            Author.findByName("name2")
-        } == null
-        EntityInSecondDatastore.withNewTransaction {
-            EntityInSecondDatastore.findByName("name2")
-        } != null
-        AuditTrail.withNewTransaction {
-            AuditTrail.list().collect { it.className }.unique()
-        } == ["test.EntityInSecondDatastore"]
-    }
-
-    void "test nested transactions different datastores: rollback inner"() {
-        expect:
-        AuditTrail.withNewTransaction {
-            AuditTrail.list().collect { it.className }.unique()
-        } == []
-
-        when:
-        Author.withNewTransaction {
-            EntityInSecondDatastore.withNewTransaction { TransactionStatus transactionStatus ->
-                new Author(name: "name2", age: 12, famous: true).save(flush: true, failOnError: true)
-                new EntityInSecondDatastore(name: "name2", someIntegerProperty: 1).save(flush: true, failOnError: true)
-                // Rollback new EntityInSecondDatastore
-                transactionStatus.setRollbackOnly()
-            }
-            // Commit new Author
-        }
-
-        then:
-        Author.withNewTransaction {
-            Author.findByName("name2")
-        } != null
-        EntityInSecondDatastore.withNewTransaction {
-            EntityInSecondDatastore.findByName("name2")
-        } == null
-        AuditTrail.withNewTransaction {
-            AuditTrail.list().collect { it.className }.unique()
-        } == ["test.Author"]
     }
 }
